Implement breadcrumbs on all pages (except the front page)
Breadcrumbs are located between the header and the main section. A list
of breadcrumb "steps" (crumbs) need to be put in PHP for Smarty to
iterate over and display. Each breadcrumb has a link, and a label (I'm
not super attached to the way it looks so I'm open to ideas). If we
forget to put a breadcrumb in the PHP a big red box will be displayed to
remind us.

This looked simple but having it display properly was tricky: I wanted
to write the code only once in `main.html` rather than having to copy it
on all pages, and yet it had to be inside the DIV with id `main` to
display properly. That led to a reshuffling of the layout.

I also had to replace the various `<br>` that were used to space tiles
with proper margins, otherwise the top tile started with a `<br>` and
the spacing with the breadcrumb was too high.

I think I fixed everything, but it's worth re-testing with multiple
screen sizes.