Organize the main templates in sub-folders
It always bothered me that the templates are all in a single folder as I
find it makes it hard to navigate. This is especially true on the CPANEL
which has a ton of template.

I decided to do a bit of spring cleanup and organize the templates by
folder. I started with the main site, and will move to the CPANEL in a
later stage.

This also has the benefit of removing the `1/...` in the template paths
so that they are not tied to the theme number 1.