Fixes #742 for good hopefully
The lightbox already existed in the list of interviews, however it did
not work because of a Javascript error caused by `comments.js`: It tried
to use jQuery which is loaded after.

The root cause is that the comment tile was including `comment.js`, and
it can be in the middle of the page, but jQuery is always included last.
So remove the script inclusion from the comment tile, and added
`comment.js` in the additional scripts of all pages that need it.