Add table needed for Laravel DB migrations
This is to prepare the transition to Laravel. Laravel uses DB migrations
to modify the database schema, very similar to the custom DB scripts we
have. The `migrations` table is similar to our `database_change` table.

The migrations table needs to be initialized with all the migrations
needed to setup the initial schema. After that we can add further
migrations to change the DB structure via Laravel.

In practice there are no impacts on the existing site for this, since it
just adds a table that no code uses.