Adjust data folder for the prototype deployment
When deploying inside the Laravel prototype the path to the data folder
is wrong, as it's not in `/data` anymore at the root of the site, but
one level above and in `storage/`.

For now create a symlink between the two, to not have to change all
paths in the legacy code.