Fixes atari-legend/atari-legend#152: Wrong ID passed to the changelog
The ID passed to the changelog function was the comment ID instead of
the game ID. That worked most of the time because the comment IDs are
in the 1000's and we do have games with IDs in 1000's, so the query
would work. The data was wrong, but it's not actually used as it's
re-queried later.

However it may fail if by chance the comment ID did not match a game ID,
which may happen as we don't have a continuous list of IDs for game but
there are some gaps.

In any case, pass the correct game ID rather than the comment ID to
prevent any issues.